requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
numpy
pillow
python-multipart
paddleocr
httpx
